pyproject.toml
[project]
name = "wikrev"
version = "0.1.2"
description = "WikRev - Wiki Reviewer: Review recent wiki changes with diffs and previews."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "jinja2>=3.1.4",
    "markdown>=3.6",
    "pydantic>=2.7.0",
    "github-copilot-sdk>=0.1.0",
]

[project.scripts]
wikrev = "wikrev.app:main"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/wikrev"]

[tool.hatch.metadata]
allow-direct-references = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
